+++ targets/TARGET_RASPBERRYPI/TARGET_MCU_RP2/i2c_api.c
@@ -56,9 +56,10 @@ static unsigned int const DEFAULT_I2C_BAUDRATE = 100 * 1000; /* 100 kHz */
 void i2c_init(i2c_t *obj, PinName sda, PinName scl)
 {
     /* Verify if both pins belong to the same I2C peripheral. */
-    I2CName const i2c_sda = (I2CName)pinmap_peripheral(sda, PinMap_I2C_SDA);
-    I2CName const i2c_scl = (I2CName)pinmap_peripheral(scl, PinMap_I2C_SCL);
-    MBED_ASSERT(i2c_sda == i2c_scl);
+    uint32_t const i2c_sda = pinmap_peripheral(sda, PinMap_I2C_SDA);
+    uint32_t const i2c_scl = pinmap_peripheral(scl, PinMap_I2C_SCL);
+    const uint32_t i2c_peripheral = pinmap_merge(i2c_sda, i2c_scl);
+    MBED_ASSERT(i2c_peripheral != ((uint32_t)NC));
 
 #if DEVICE_I2CSLAVE
     /** was_slave is used to decide which driver call we need
@@ -87,6 +88,10 @@ void i2c_init(i2c_t *obj, PinName sda, PinName scl)
     gpio_pull_up(scl);
 }
 
+void i2c_free(i2c_t *obj) {
+    i2c_deinit(obj->i2c.dev);
+}
+
 void i2c_frequency(i2c_t *obj, int hz)
 {
     DEBUG_PRINTF("obj->i2c.is_slave: %d\r\n", obj->i2c.is_slave);
@@ -100,8 +105,114 @@ void i2c_frequency(i2c_t *obj, int hz)
     obj->i2c.baudrate = i2c_set_baudrate(obj->i2c.dev, hz);
 }
 
+int i2c_start(i2c_t *obj) {
+    (void)obj;
+    // No-op, we need to wait until we get the address to start
+    return 0;
+}
+
+int i2c_byte_write(i2c_t *obj, int data) {
+    int ret = 1;
+
+    if(obj->state == MBED_HAL_I2C_STATE_STARTED) {
+        // First byte (address). Save for now in TAR register (won't actually get sent until later)
+        obj->i2c.dev->hw->enable = 0;
+        obj->i2c.dev->hw->tar = data >> 1;
+        obj->i2c.dev->hw->enable = 1;
+    }
+    else {
+        // Write byte. Force transmission of a start condition if this is the first data byte.
+        obj->i2c.dev->hw->data_cmd = data | (obj->state == MBED_HAL_I2C_STATE_ADDRESSED ? I2C_IC_DATA_CMD_RESTART_BITS : 0);
+
+        // Wait until the transmission of the address/data from the internal
+        // shift register has completed. For this to function correctly, the
+        // TX_EMPTY_CTRL flag in IC_CON must be set. The TX_EMPTY_CTRL flag
+        // was set in pico_sdk_i2c_init.
+        do {
+            tight_loop_contents();
+        }
+        while (!(obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_TX_EMPTY_BITS));
+
+        // Did we encounter an error?
+        if(obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_TX_ABRT_BITS) {
+            // Get the abort reason
+            uint32_t abort_reason = obj->i2c.dev->hw->tx_abrt_source;
+            if(abort_reason & (I2C_IC_TX_ABRT_SOURCE_ABRT_TXDATA_NOACK_BITS | I2C_IC_TX_ABRT_SOURCE_ABRT_10ADDR1_NOACK_BITS | I2C_IC_TX_ABRT_SOURCE_ABRT_10ADDR2_NOACK_BITS | I2C_IC_TX_ABRT_SOURCE_ABRT_7B_ADDR_NOACK_BITS)) {
+                ret = 0; // no ACK
+            }
+            else {
+                ret = 3; // other error
+            }
+
+            // Don't clear the abort yet as we check this in i2c_stop()
+        }
+    }
+
+    return ret;
+}
+
+int i2c_byte_read(i2c_t *obj, int last) {
+
+    // This I2C peripheral waits to generate the ACK/NACK after a read until either we try to read the
+    // next byte or we generate a stop/repeated start. This means it automatically uses the correct
+    // type of ACK or NACK and doesn't need (and can't accept) a flag on whether to do so.
+    (void)last;
+
+    // Trigger reading of a byte. Force generation of a start condition if this is the first byte.
+    obj->i2c.dev->hw->data_cmd = I2C_IC_DATA_CMD_CMD_BITS |
+        (obj->state == MBED_HAL_I2C_STATE_ADDRESSED ? I2C_IC_DATA_CMD_RESTART_BITS : 0);
+
+    // Wait until the transmission of the address/data from the internal
+    // shift register has completed. For this to function correctly, the
+    // TX_EMPTY_CTRL flag in IC_CON must be set. The TX_EMPTY_CTRL flag
+    // was set in pico_sdk_i2c_init.
+    do {
+        tight_loop_contents();
+    }
+    while (!(obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_TX_ABRT_BITS) && !i2c_get_read_available(obj->i2c.dev));
+
+    // Did we encounter an error?
+    if(obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_TX_ABRT_BITS) {
+        // Sadly no way to pass up the abort reason so no reason to read tx_abrt_source.
+        // Also don't clear the abort yet as we check this in i2c_stop()
+        return 0;
+    }
+
+    // Return data
+    return (uint8_t) obj->i2c.dev->hw->data_cmd;
+}
+
+int i2c_stop(i2c_t *obj)
+{
+    // If we didn't already generate a stop due to an error earlier in the transaction...
+    if(!(obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_TX_ABRT_BITS)) {
+        // ...then generate one now.
+        obj->i2c.dev->hw->enable |= I2C_IC_ENABLE_ABORT_BITS;
+        do {
+            tight_loop_contents();
+        }
+        while (!(obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_TX_ABRT_BITS));
+    }
+
+    // Clear stop detection and abort reason flags (clear on read registers, ick!).
+    // If we don't clear these the SDK read/write functions will get stuck.
+    obj->i2c.dev->hw->clr_tx_abrt;
+    obj->i2c.dev->hw->clr_stop_det;
+
+    return 0;
+}
+
 int i2c_read(i2c_t *obj, int address, char *data, int length, int stop)
 {
+    // Make sure a repeated start is correctly generated if we currently have the bus locked
+    obj->i2c.dev->restart_on_next = true;
+
+    // If we were previously doing a single byte transaction which aborted for any reason,
+    // ensure we clear that abort so that the Pico SDK function doesn't get stuck
+    if(obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_TX_ABRT_BITS) {
+        i2c_stop(obj);
+    }
+
     int const bytes_read = i2c_read_blocking(obj->i2c.dev,
                                              (uint8_t)(address >> 1),
                                              (uint8_t *)data,
@@ -115,13 +226,13 @@ int i2c_read(i2c_t *obj, int address, char *data, int length, int stop)
 
 int i2c_write(i2c_t *obj, int address, const char *data, int length, int stop)
 {
-    if (length == 0) {
-        // From pico-sdk:
-        // static int i2c_write_blocking_internal(i2c_inst_t *i2c, uint8_t addr, const uint8_t *src, size_t len, bool nostop,
-        // Synopsys hw accepts start/stop flags alongside data items in the same
-        // FIFO word, so no 0 byte transfers.
-        // invalid_params_if(I2C, len == 0);
-        length = 1;
+    // Make sure a repeated start is correctly generated if we currently have the bus locked
+    obj->i2c.dev->restart_on_next = true;
+
+    // If we were previously doing a single byte transaction which aborted for any reason,
+    // ensure we clear that abort so that the Pico SDK function doesn't get stuck
+    if(obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_TX_ABRT_BITS) {
+        i2c_stop(obj);
     }
 
     int const bytes_written = i2c_write_blocking(obj->i2c.dev,
@@ -161,9 +272,17 @@ const PinMap *i2c_slave_scl_pinmap()
     return PinMap_I2C_SCL;
 }
 
-int i2c_stop(i2c_t *obj)
+// Report I2C capabilities
+static const i2c_capabilities_t i2c_caps = {
+    .single_byte_address_delayed = true,
+    .single_byte_start_cond_delayed = true,
+    .supports_single_byte = true,
+    .supports_zero_length_transfer_single_byte = false,
+    .supports_zero_length_transfer_transaction = false
+};
+MBED_WEAK i2c_capabilities_t const * i2c_get_capabilities()
 {
-    return 0;
+    return &i2c_caps;
 }
 
 #if DEVICE_I2CSLAVE
@@ -206,54 +325,99 @@ int i2c_slave_receive(i2c_t *obj)
     return (retValue);
 }
 
-/** Configure I2C as slave or master.
- *  @param obj The I2C objecti2c_get_read_availableread
- *  @return non-zero if a value is available
- */
 int i2c_slave_read(i2c_t *obj, char *data, int length)
 {
     int bytes_read = 0;
-    for (size_t i = 0; i < (size_t)length; ++i) {
-        while (!i2c_get_read_available(obj->i2c.dev)) {
+
+    while(true) {
+        // Wait until something happens
+        while (!i2c_get_read_available(obj->i2c.dev) && !(obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_STOP_DET_BITS)) {
             tight_loop_contents();
         }
 
-        *data = obj->i2c.dev->hw->data_cmd;
-        bytes_read++;
+        // Drain Rx FIFO while there is data
+        while(i2c_get_read_available(obj->i2c.dev)) {
+            if(bytes_read < length) {
+                data[bytes_read++] = obj->i2c.dev->hw->data_cmd;
+            }
+            else {
+                obj->i2c.dev->hw->data_cmd; // throw away data
+            }
+        }
 
-        // Check stop condition
-        bool stop = (obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_STOP_DET_BITS) != 0;
-        if (stop && !i2c_get_read_available(obj->i2c.dev)) {
+        // If we have received a stop, then bail
+        if((obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_STOP_DET_BITS)) {
+            // Race condition: did we get any data after exiting the while loop above? If so then loop again.
+            if(i2c_get_read_available(obj->i2c.dev)) {
+                continue;
+            }
+
             // Clear stop (by reading the register)
-            int clear_stop = obj->i2c.dev->hw->clr_stop_det;
-            (void)clear_stop;
+            obj->i2c.dev->hw->clr_stop_det;
+
             break;
-        } else {
-            data++;
         }
     }
 
     return bytes_read;
 }
 
-/** Configure I2C as slave or master.
- *  @param obj The I2C object
- *  @param data    The buffer for sending
- *  @param length  Number of bytes to write
- *  @return non-zero if a value is available
- */
 int i2c_slave_write(i2c_t *obj, const char *data, int length)
 {
     DEBUG_PRINTF("i2c_slave_write\r\n");
 
-    i2c_write_raw_blocking(obj->i2c.dev, (const uint8_t *)data, (size_t)length);
+    int bytes_written = 0;
 
-    // Clear interrupt (by reading the register)
-    int clear_read_req = i2c_get_hw(obj->i2c.dev)->clr_rd_req;
-    (void)clear_read_req;
-    DEBUG_PRINTF("clear_read_req: %d\n", clear_read_req);
+    while(true) {
+        // Wait for something to happen
+        while (!i2c_get_write_available(obj->i2c.dev) && 
+            !(obj->i2c.dev->hw->raw_intr_stat & (I2C_IC_RAW_INTR_STAT_TX_ABRT_BITS | I2C_IC_RAW_INTR_STAT_STOP_DET_BITS | I2C_IC_RAW_INTR_STAT_RD_REQ_BITS))) {
+            tight_loop_contents();
+        }
 
-    return length;
+        if(obj->i2c.dev->hw->raw_intr_stat & (I2C_IC_RAW_INTR_STAT_TX_ABRT_BITS | I2C_IC_RAW_INTR_STAT_STOP_DET_BITS)) {
+            // Transaction ended
+            break;
+        }
+
+        // Feed more bytes into the FIFO if we have them and there's room
+        if(i2c_get_write_available(obj->i2c.dev) && bytes_written < length) {
+            obj->i2c.dev->hw->data_cmd = data[bytes_written++];
+
+            // Tell the hardware we gave it some data
+            obj->i2c.dev->hw->clr_rd_req;
+        }
+
+        if((obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_RD_REQ_BITS) && (bytes_written >= length)) {
+            // Out of data, and currently stretching the clock.
+            // To un-stick the master, for now we will just feed zeros into the FIFO.
+            // This error case could potentially be improved later.
+            obj->i2c.dev->hw->data_cmd = 0;
+            obj->i2c.dev->hw->clr_rd_req;
+        }
+    }
+
+    // Clear stop flag always
+    obj->i2c.dev->hw->clr_stop_det;
+
+    // Handle success vs failure
+    if(obj->i2c.dev->hw->raw_intr_stat & I2C_IC_RAW_INTR_STAT_TX_ABRT_BITS) {
+        if(obj->i2c.dev->hw->tx_abrt_source & I2C_IC_TX_ABRT_SOURCE_ABRT_SLVFLUSH_TXFIFO_BITS) {
+            // Master ended transaction early.
+            // Thankfully there is a useful field that shows the number of bytes
+            // written to the FIFO that were flushed due to transaction end.
+            const int bytes_actually_written = bytes_written - (obj->i2c.dev->hw->tx_abrt_source >> I2C_IC_TX_ABRT_SOURCE_TX_FLUSH_CNT_LSB);
+            obj->i2c.dev->hw->clr_tx_abrt;
+            return bytes_actually_written;
+        }
+        else {
+            // Other error
+            obj->i2c.dev->hw->clr_tx_abrt;
+            return -1;
+        }
+    }
+
+    return bytes_written;
 }
 
 /** Configure I2C address.